+++ maxpool.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 using namespace std;
 
-double* maxpool(int filter_size,int matrix_size,int stride,double* A){
+float* maxpool(int filter_size,int matrix_size,int stride,float* A){
 // considering only the case when filter_size divides the matrix_size
 // output_size-1<=(matrix_size-filter_size)/stride
   int output_size = (matrix_size+stride-1)/stride;
 
-  double * C = new double[output_size*output_size];
+  float * C = new float[output_size*output_size];
 
 
   int i = 0;
@@ -38,10 +38,10 @@ double* maxpool(int filter_size,int matrix_size,int stride,double* A){
  return C;
 }
 
-double* maxpool(int filter_size,int matrix_size,int stride,int padding_size,double* A){
+float* maxpool(int filter_size,int matrix_size,int stride,int padding_size,float* A){
   int new_matrix_size = matrix_size + 2*padding_size;
 
-  double * B = new double[new_matrix_size*new_matrix_size];
+  float * B = new float[new_matrix_size*new_matrix_size];
 
 
   for(int i = 0;i<new_matrix_size;i++){
@@ -57,7 +57,7 @@ double* maxpool(int filter_size,int matrix_size,int stride,int padding_size,doub
     }
   }
 
-  double * C = maxpool(filter_size,new_matrix_size,stride,B);
+  float * C = maxpool(filter_size,new_matrix_size,stride,B);
   return C;
 }
 
@@ -67,9 +67,9 @@ double* maxpool(int filter_size,int matrix_size,int stride,int padding_size,doub
 int main(){
   int matrix_size = 10; int filter_size = 3; int stride=2;
   int output_size = (matrix_size-filter_size+stride)/stride;
-  double** a = new double*[matrix_size];
+  float** a = new float*[matrix_size];
   for(int i = 0;i<matrix_size;i++){
-    a[i]= new double[matrix_size];
+    a[i]= new float[matrix_size];
   }
   for(int i = 0;i<matrix_size;i++){
     for(int j = 0;j<matrix_size;j++ ){
@@ -77,7 +77,7 @@ int main(){
     }
 
   }
-  double **b ;
+  float **b ;
   b = maxpool(filter_size,matrix_size,stride,a);
 
   for(int i = 0;i<output_size;i++){
